backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "MindWise - AI-powered job application decision-support system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pydantic[email]>=2.0.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "pdfplumber>=0.11.0",
    "google-generativeai>=0.3.0",
    "google-genai>=1.59.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
]
